Fix import paths after file reorganization
Update import statements to use new file location:
- openhcs.ui.shared.pyqt6_widget_strategies → openhcs.pyqt_gui.widgets.shared.widget_strategies

Fixed imports in:
- openhcs/pyqt_gui/dialogs/group_by_selector_dialog.py (2 imports)
- openhcs/pyqt_gui/widgets/function_list_editor.py (1 import)

This resolves the ModuleNotFoundError that occurred after the merge
due to file reorganization in the ui-refactor branch.
