+++ openhcs/pyqt_gui/dialogs/group_by_selector_dialog.py
@@ -58,7 +58,7 @@ class GroupBySelectorDialog(QDialog):
     
     def setup_ui(self):
         """Setup the user interface (mirrors Textual TUI layout)."""
-        from openhcs.ui.shared.pyqt6_widget_strategies import _get_enum_display_text
+        from openhcs.pyqt_gui.widgets.shared.widget_strategies import _get_enum_display_text
         component_display = _get_enum_display_text(self.group_by).title()
 
         self.setWindowTitle(f"Select {component_display}s")
@@ -191,7 +191,7 @@ class GroupBySelectorDialog(QDialog):
         Returns:
             Formatted display string (e.g., "Channel 1 | HOECHST 33342" or "Channel 1")
         """
-        from openhcs.ui.shared.pyqt6_widget_strategies import _get_enum_display_text
+        from openhcs.pyqt_gui.widgets.shared.widget_strategies import _get_enum_display_text
         component_display = _get_enum_display_text(self.group_by).title()
         base_text = f"{component_display} {component_key}"
